requirements.txt
flask
flask-cors
urllib3
PyYAML
PyMySQL
DBUtils
github-flask